requirements.txt
pdfplumber==0.11.0
chromadb==0.5.23
openai>=2.0.0
streamlit==1.39.0
python-dotenv==1.0.1
loguru==0.7.2
pandas==2.2.3
pymysql>=1.1.0
